DEVOPS-8307 refactor provider structure to the current scaffolding schema
